Chat: shared theme toggle, timestamps, join/leave notices
- Theme: extract the color-scheme theming from app.ts into a shared theme.ts and
  import it in both app.ts and chat.ts, so the shmee-man button cycles themes on
  the chat page exactly as on the other pages (same schemes, same apply logic).
- Timestamps + names: the server now stamps every line "[time] name: text". The
  client passes its name in the join URL and sends raw text; the server formats
  it, so the time and name are consistent and not spoofable per message.
- Status / server updates: joining a room broadcasts "* name joined *" and
  leaving broadcasts "* name left *" (system messages, broadcast but not
  persisted, so they stay out of replayed history). The client shows a local
  "disconnected" status on close.

Tests updated for the server-stamped format (scan frames past join notices);
store/history/unknown-room coverage unchanged.
