+++ site/static/css/chat.css
@@ -53,6 +53,12 @@
     word-break: break-word;
 }
 
+.statusLine {
+    padding: 1px 0;
+    font-style: italic;
+    opacity: 0.7;
+}
+
 #send_msg_form {
     display: flex;
     gap: 8px;
